Implemented creating, updating, and deleting pending operations.

+++ HealthVault/Classes/Transport/MHVRequestMessageCreator.m
@@ -111,7 +111,7 @@
 {
     [header appendXmlElement:@"method" text:self.method.name];
     [header appendXmlElementStart:@"method-version"];
-    [header appendFormat:@"%.0d", self.method.version];
+    [header appendFormat:@"%.0ld", (long)self.method.version];
     [header appendXmlElementEnd:@"method-version"];
 }
